requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Machine learning
scikit-learn>=1.0.0
joblib>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Optional enhanced functionality
umap-learn>=0.5.0
lightgbm>=3.2.0
networkx>=2.6.0
pywavelets>=1.1.0

# Deep learning (optional)
torch>=1.9.0; platform_system != "Darwin" or platform_machine != "arm64"
captum>=0.5.0; platform_system != "Darwin" or platform_machine != "arm64"

# Statistical analysis
statsmodels>=0.12.0

# Topology analysis (optional)
ripser>=0.6.0

# Hyperparameter optimization (optional)
optuna>=2.10.0

# Progress bars
tqdm>=4.60.0

# Configuration
pyyaml>=5.4.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0 